# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText
  description: |-
    The user is presented with a list of conversations with other users or with groups, sorted in reverse
    chronological order. Each element in the list must display the username of the other person or the
    group name, the user profile photo or the group photo, the date and time of the latest message,
    the preview (snippet) of the text message, or an icon for a photo message. The user can start a new
    conversation with any other user of WASAText, and this conversation will automatically be added
    to the list. The user can search for other users via the username and see all the existing WASAText
    usernames.
    The user can create a new group with any number of other WASAText users to start a conversation.
    Group members can add other users to the group, but users cannot join groups on their own or even
    see groups they aren’t a part of. Additionally, users have the option to leave a group at any time.
    The user can open a conversation to view all exchanged messages, displayed in reverse chronological
    order. Each message includes the timestamp, the content (whether text or photo), and the sender’s
    username for received messages, or one/two checkmarks to indicate the status of sent messages. Any
    reactions (comments) on messages are also displayed, along with the names of the users who posted
    them.
...
    recipient (by all the recipients for groups) within the conversation itself.
...
servers:
...
paths:
  /session:
    post:
      tags: ["login"]
      summary: Logs in the user
      description: |-
        If the user does not exist, it will be created,
        and an identifier is returned.
        If the user exists, the user identifier is returned.
      operationId: doLogin
      requestBody:
        description: Login request body
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Object containing username for login.
              properties:
                username:
                  $ref: "#/components/schemas/Username"
              required:
                - username
      responses:
        '201':
          description: User log-in action successful
          content:
            application/json:
              schema:
                type: object
                description: id and photo of the user logged in.
                properties:
                  userid:
                    $ref: "#/components/schemas/UserId"
                  photo:
                    $ref: "#/components/schemas/Photo"
                required:
                  - userid
                  - photo
              example:
                userid: 1
                photo: "https://example.com/photo.jpg"
        '400':
          $ref: "#/components/responses/BadRequest"
        '500':
          $ref: "#/components/responses/InternalServerError"
...
  /chats/{chat_id}/messages:
    parameters:
      - $ref: "#/components/parameters/ChatId"
    post:
      security:
        - bearerAuth: []
      tags: ["message"]
      summary: Sends a message to the chat
      description: |-
        Sends a message to the chat associated with the chat id.
      operationId: sendMessage
      requestBody:
        description: Message to send.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageToSend"
      responses:
        '201':
          description: Message sent successfully
          content:
            application/json:
              schema:
                type: object
                description: Message id response.
                properties:
                  messageid:
                    $ref: "#/components/schemas/MessageId"
                required:
                  - messageid
              example:
                messageid: 100
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
  /chats/{chat_id}/forwardedmessages:
    parameters:
      - $ref: "#/components/parameters/ChatId"
    post:
      security:
        - bearerAuth: []
      tags: ["message"]
      summary: Forwards a message to the chat
      description: |-
        Forwards a message to the chat associated with the chat id.
      operationId: forwardMessage
      requestBody:
        description: Message to forward.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageToSend"
      responses:
        '201':
          description: Message forwarded successfully
          content:
            application/json:
              schema:
                type: object
                description: Message id response.
                properties:
                  messageid:
                    $ref: "#/components/schemas/MessageId"
                required:
                  - messageid
              example:
                messageid: 101
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
  /chats/{chat_id}/repliedmessages:
    parameters:
      - $ref: "#/components/parameters/ChatId"
    post:
      security:
        - bearerAuth: []
      tags: ["message"]
      summary: Replies to a message in the chat
      description: |-
        Replies to a message in the chat associated with the chat id.
      operationId: replyMessage
      requestBody:
        description: Message to reply.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageReply"
      responses:
        '201':
          description: Message replied successfully
          content:
            application/json:
              schema:
                type: object
                description: Message id response.
                properties:
                  messageid:
                    $ref: "#/components/schemas/MessageId"
                required:
                  - messageid
              example:
                messageid: 102
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
  /chats/{chat_id}/messages/{message_id}:
    parameters:
      - $ref: "#/components/parameters/ChatId"
      - $ref: "#/components/parameters/MessageId"
    delete:
      security:
        - bearerAuth: []
      tags: ["message"]
      summary: Deletes a message in the chat
      description: |-
        Deletes a message in the chat associated with the chat id.
      operationId: deleteMessage
      responses:
        '200':
          description: Message deleted successfully
          content:
            application/json:
              schema:
                type: object
                description: Confirmation message.
                properties:
                  message:
                    type: string
                    description: Confirmation message.
                    pattern: "^.*$"
                    minLength: 1
                    maxLength: 999
              example:
                message: "Message deleted"
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
  /chats/{chat_id}/messages/{message_id}/comments:
    parameters:
      - $ref: "#/components/parameters/ChatId"
      - $ref: "#/components/parameters/MessageId"
    put:
      security:
        - bearerAuth: []
      tags: ["message"]
      summary: Comments a message in the chat
      description: |-
        Comments a message in the chat associated with the chat id.
      operationId: commentMessage
      requestBody:
        description: Comment to the message.
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Emoji comment object.
              properties:
                emoji:
                  $ref: "#/components/schemas/Emoji"
              required:
                - emoji
      responses:
        '200':
          description: Message commented successfully
          content:
            application/json:
              schema:
                type: object
                description: Confirmation message.
                properties:
                  message:
                    type: string
                    description: Confirmation message.
                    pattern: "^.*$"
                    minLength: 1
                    maxLength: 999
              example:
                message: "Message commented"
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
    delete:
      security:
        - bearerAuth: []
      tags: ["message"]
      summary: Uncomments a message in the chat
      description: |-
        Uncomments a message in the chat associated with the chat id.
      operationId: uncommentMessage
      responses:
        '200':
          description: Message uncommented successfully
          content:
            application/json:
              schema:
                type: object
                description: Confirmation message.
                properties:
                  message:
                    type: string
                    description: Confirmation message.
                    pattern: "^.*$"
                    minLength: 1
                    maxLength: 999
              example:
                message: "Message uncommented"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
components:
  parameters:
    ChatId:
      name: chat_id
      in: path
      description: ID of the chat.
      required: true
      schema:
        type: integer
        minimum: 1
        maximum: 2147483647
        example: 1
        readOnly: true
...
    MessageId:
      name: message_id
      in: path
      description: ID of the message.
      required: true
      schema:
        type: integer
        minimum: 1
        maximum: 2147483647
        example: 1
        readOnly: true
...
  schemas:
    Username:
      title: username
      description: Username of the user.
      type: string
      example: "user1"
      pattern: "^.*$"
      minLength: 3
      maxLength: 16
...
    Photo:
      title: photo
      description: Photo of the user or group.
      type: string
      format: url
      example: "https://example.com/photo.jpg"
      pattern: "^.*$"
      minLength: 1
      maxLength: 999
    UserId:
      title: userid
      description: ID of the user.
      type: integer
      minimum: 1
      maximum: 2147483647
      example: 1
...
    MessageId:
      title: messageid
      description: ID of the message.
      type: integer
      minimum: 1
      maximum: 2147483647
      example: 1
...
    TextData:
      title: textdata
      description: Text of the message.
      type: string
      example: "Hello, how are you?"
      pattern: "^(.|\\n)*$"
      minLength: 1
      maxLength: 999
    Emoji:
      title: emoji
      description: Emoji to add as comment to the message.
      type: string
      example: "😊"
      pattern: "^😊$"
      minLength: 1
      maxLength: 1
...
    MessageToSend:
      title: messagetosend
      description: Message to send.
      type: object
      properties:
        text:
          $ref: "#/components/schemas/TextData"
        photo:
          $ref: "#/components/schemas/Photo"
      required:
        - text
    MessageReply:
      title: messagereply
      description: Message to reply.
      type: object
      properties:
        text:
          $ref: "#/components/schemas/TextData"
        photo:
          $ref: "#/components/schemas/Photo"
        repliedMessageId:
          $ref: "#/components/schemas/MessageId"
      required:
        - text
        - repliedMessageId
...
  responses:
    BadRequest:
      description: The request was not compliant with the documentation.
      content:
        application/json:
          schema:
            type: object
            description: Error response object.
            properties:
              error:
                type: string
                description: Error message.
                pattern: "^.*$"
                minLength: 1
                maxLength: 200
          example:
            error: "Bad Request"
    Unauthorized:
      description: Access token is missing or invalid.
      content:
        application/json:
          schema:
            type: object
            description: Error response object.
            properties:
              error:
                type: string
                description: Error message.
                pattern: "^.*$"
                minLength: 1
                maxLength: 200
          example:
            error: "Unauthorized"
    NotFound:
      description: The server cannot find the requested resource.
      content:
        application/json:
          schema:
            type: object
            description: Error response object.
            properties:
              error:
                type: string
                description: Error message.
                pattern: "^.*$"
                minLength: 1
                maxLength: 200
          example:
            error: "Not Found"
    InternalServerError:
      description: The server encountered an unexpected condition that prevented it from fulfilling the request.
      content:
        application/json:
          schema:
            type: object
            description: Error response object.
            properties:
              error:
                type: string
                description: Error message.
                pattern: "^.*$"
                minLength: 1
                maxLength: 200
          example:
            error: "Internal Server Error"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer